server: move externalStorageBuilder out of server.go
Release note: None
