+++ pkg/server/BUILD.bazel
@@ -21,6 +21,7 @@ go_library(
         "doc.go",
         "drain.go",
         "env_sampler.go",
+        "external_storage_builder.go",
         "grpc_server.go",
         "gzip_response_writer.go",
         "import_ts.go",
